simple changes made to index.html
index.html was  changed by creating a div in the top of the grid container, and extra space for the equations once the script gets added. 
This will make the grid container look more like a calculator; I also changed the color of the opening div item0 to white within the grid-container.